dataswati/requirements.txt
# local package
-e .

# external requirements
bottleneck==1.3.2
ipywidgets==7.6.3
pandas-profiling==3.0.0
interpret==0.2.4
lightgbm==3.2.1
scikit-learn==0.23.2
fire

#dev
Sphinx
flake8
black
mypy
